Make it compile after the move
